NNS1-2851: Move wallet-account services
